+++ src/ChatVox/Settings/StartupVisibilityPolicy.cs
@@ -1,0 +1,6 @@
+namespace ChatVox.Settings;
+
+public static class StartupVisibilityPolicy
+{
+    public static bool ShouldStartHidden(AppSettings settings) => settings.StartMinimizedToTray && settings.StartMinimizedWasExplicitlySet;
+}